services/auth-svc/requirements.txt
# DB / Migrations
SQLAlchemy==2.0.32
alembic==1.13.2
python-dotenv==1.0.1
psycopg[binary]==3.2.1
psycopg2-binary==2.9.9

# App (faltaban estas)
pydantic==2.8.2
pydantic-settings==2.4.0
fastapi==0.112.2
uvicorn[standard]==0.30.6
asyncpg==0.29.0

argon2-cffi==23.1.0
email-validator==2.2.0
PyJWT==2.9.0

# Logging estructurado
python-json-logger>=2.0.7

# Redis
redis>=5.0.0

# Kafka
aiokafka>=0.10.0

# 2FA TOTP
pyotp==2.9.0
qrcode[pil]==7.4.2
